# Aintution Prospect 18-6-2026/lib/api-zod/src/generated/types/bulkProspectInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Aintution Prospects API
 * OpenAPI spec version: 0.1.0
 */
import type { ProspectInput } from './prospectInput';

export interface BulkProspectInput {
  rows: ProspectInput[];
}